requirements.txt
google-generativeai
google-genai 
langchain-google-genai
python-dotenv
streamlit
watchdog